keep dynusT output files every 10th iteration
git-svn-id: file:///srv/local/data/cleanRepo/trunk@21308 a0517364-e873-4167-b1c8-416daa5a6b91

+++ playgrounds/mrieser/src/main/java/playground/mrieser/svi/controller/DynusTControlerListener.java
@@ -127,13 +127,15 @@ public class DynusTControlerListener implements StartupListener, IterationStarts
 					event.getControler().getControlerIO().getIterationFilename(event.getIteration(), "ptStats.txt"),
 					this.dc.getTravelTimeCalculator());
 		}
-		// clean up tmp directory
-		log.info("removing unnecessary files.");
-		File outDir = new File(this.dc.getOutputDirectory());
-		for (File f : outDir.listFiles()) {
-			if (f.isFile() && !f.getName().equals("VehTrajectory.dat")) {
-				log.info("   deleting file " + f.getAbsolutePath());
-				f.delete();
+		if (event.getIteration() % 10 != 0) {
+			// clean up tmp directory
+			log.info("removing unnecessary files.");
+			File outDir = new File(this.dc.getOutputDirectory());
+			for (File f : outDir.listFiles()) {
+				if (f.isFile() && !f.getName().equals("VehTrajectory.dat")) {
+					log.info("   deleting file " + f.getAbsolutePath());
+					f.delete();
+				}
 			}
 		}
 	}
